chore: tableRowHeader의 util 함수 TableRowHeader component 파일로 이동
